Make zsh startup 100ms faster by tuning rbenv
This changes the behaviour of the rbenv initialization by skipping the
`rbenv rehash` call hidden behind `rbenv init`.

The downside to this is that rbenv won't rehash automatically when
starting a new shell instance. Which means that installing a new Ruby
version needs a manual `rbenv rehash`.

This is taken from here:

    http://kleinmann.org/2015/01/18/decreasing-my-zsh-startup-time/