Fix pretty_git_log for git >= 2.14.2
Since git >= 2.14.2 the coloring was off if the output of `git log` was
not a TTY.

See here: https://github.com/git/git/commit/136c8c8b8fa39f1315713248473dececf20f8fe7

Using `--color` fixed the problem of color disappearing when piping `git
log` to `less` in `pretty_git_log`. BUT it also colored the graph-lines
of `git log`, which in turn broke the usage of `column` in
`pretty_git_log`, since now we have ASCII color codes surrounding the
graph...

This uses `%C(always,<color>)` in the format for the log message to
always enable these colors -- no matter if it's a TTY or not we're
outputting to. And it leaves the graph-lines without color.